Move pricing menu to own component
