+++ packages/js/product-editor/src/components/variations-table/pricing-menu-item/index.ts
@@ -1,0 +1,1 @@
+export * from './pricing-menu-item';
